batch/requirements.txt
pandas
openpyxl
tqdm
argparse
python-dotenv